remove bears works
