+++ styles/main.css
@@ -7,3 +7,7 @@ input{
   background-color: white !important;
   border-radius: 5px;
 }
+
+.close {
+  text-align: right;
+}
